refactored the complete visualization
